Refactor dataset paths in `prep_data.py` and `vector_loader.py`, enhance SF Public Data loading with improved error handling and logging, update JSON schemas for multiple datasets, and remove obsolete `analysis_map.json`. Adjust aggregation functions in `genAggregate.py` and `genChart.py` to use 'mean' for specific fields. Update dataset queries in JSON files for better accuracy.

+++ ai/tools/anomaly_detection.py
@@ -28,7 +28,7 @@ def calculate_stats(values):
     mean = sum(values) / n
     variance = sum((x - mean) ** 2 for x in values) / n
     std_dev = variance ** 0.5
-    return {'mean': int(mean), 'stdDev': int(std_dev)}
+    return {'mean': float(mean), 'stdDev': float(std_dev)}
 
 def get_date_ranges():
     from datetime import date, timedelta
